+++ src/main/java/com/example/gamemate/domain/review/dto/ReviewUpdateResponseDto.java
@@ -1,25 +1,25 @@
-package com.example.gamemate.review.dto;
+package com.example.gamemate.domain.review.dto;
 
-import com.example.gamemate.review.entity.Review;
+import com.example.gamemate.domain.review.entity.Review;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
 
 @Getter
-public class ReviewCreateResponseDto {
+public class ReviewUpdateResponseDto {
     private  Long id;
     private  String content;
     private  Integer star;
     private  Long gameId;
     private  Long userId;
-    private  LocalDateTime createdAt;
+    private  LocalDateTime modifiedAt;
 
-    public ReviewCreateResponseDto(Review review) {
+    public ReviewUpdateResponseDto(Review review) {
         this.id = review.getId();
         this.content = review.getContent();
         this.star = review.getStar();
         this.gameId = review.getGame().getId();
         this.userId = review.getUserId();
-        this.createdAt = review.getCreatedAt();
+        this.modifiedAt = review.getModifiedAt();
     }
 }
